+++ ExistingStatesTree.h
@@ -5,7 +5,7 @@
 #include "State.h"
 
 struct Node {
-	Node(uint16_t number);
+	Node(uint8_t number);
 	uint8_t number;
 	uint16_t current_moves; // 0 if it is not on the 16 level
 	std::vector<Node*> nextNumber;
